requirements.txt
apt-clone==0.2.1
apt-xapian-index==0.47
apturl==0.5.2
beautifulsoup4==4.5.3
blinker==1.3
Brlapi==0.6.4
chardet==2.3.0
command-not-found==0.3
cryptography==1.2.3
defer==1.0.6
dirspec==13.10
Django==1.10.5
django-geojson==2.9.1
django-leaflet==0.20.0
djangorestframework==3.5.4
geoip2==2.4.2
httplib2==0.9.1
idna==2.0
jsonfield==1.0.3
louis==2.6.4
lxml==3.5.0
maxminddb==1.2.3
oauthlib==1.0.3
oneconf==0.3.9
pexpect==4.0.1
Pillow==3.1.2
piston-mini-client==0.7.5
ptyprocess==0.5
pyasn1==0.1.9
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyICU==1.9.2
PyJWT==1.3.0
python-apt==1.1.0b1
python-debian==0.1.27
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
sessioninstaller==0.0.0
six==1.10.0
system-service==0.3
thin-client-config-agent==0.8
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
virtkey==0.63.0
wheel==0.24.0
xkit==0.0.0
